Crate net_traits

Modules

blob_url_store

fetch

An implementation of the Fetch specification

filemanager_thread

http_status

image_cache

mime_classifier

policy_container

pub_domains

Implementation of public domain matching.

quality

request

resource_fetch_timing 🔒

response

The Response object resulting from a fetch operation

Structs

BlobTokenRefreshRequest

BlobTokenRevocationRequest

CacheEntryDescriptor

CookieAsyncResponse

CookieChange

CookieOperationId

Identifies a pending asynchronous cookie operation initiated by the embedder.

CustomResponse

CustomResponseMediator

DiscardFetch

A fetch task that discards all data it’s sent, useful when speculatively prefetching data that we don’t need right now, but might need in the future.

FetchThread

A thread to handle fetches in a Servo process. This thread is responsible for listening for new fetch requests as well as updates on those operations and forwarding them to crossbeam channels.

FetchThreadHandle 🔒

Metadata

Metadata about a loaded resource, such as is obtained from HTTP headers.

ResourceCorsData

ResourceFetchTiming

ResourceFetchTimingContainer

A simple container of ResourceFetchTiming to allow easy sharing between threads and allow to set multiple attributes in sequence.

ResourceThreads

SiteDescriptor

TlsSecurityInfo

Enums

CookieData

CookieSource

The creator of a given cookie

CoreResourceMsg

FetchChannels

IPC channels to communicate with the script thread about network or DOM events.

FetchMetadata

FetchResponseMsg

FilteredMetadata

IncludeSubdomains

LoadContext

A loading context, for context-specific sniffing, as defined in https://mimesniff.spec.whatwg.org/#context-specific-sniffing

MessageData

NetworkError

Network errors that have to be exported out of the loaders

RedirectEndValue

RedirectStartValue

ReferrerPolicy

Policies for providing a referrer header for a request

ResourceAttribute

ResourceTimeValue

ResourceTimingType

TlsSecurityState

ToFetchThreadMessage 🔒

WebSocketDomAction

WebSocketNetworkEvent

Constants

DOCUMENT_ACCEPT_HEADER_VALUE

https://fetch.spec.whatwg.org/#document-accept-header-value

Statics

FETCH_THREAD 🔒

PRIVILEGED_SECRET

Traits

AsyncRuntime

Handle to an async runtime, only used to shut it down for now.

FetchTaskTarget

Functions

cancel_async_fetch

Instruct the resource thread to cancel an existing request. Does nothing if the request has already completed or has not been fetched yet.

fetch_async

Instruct the fetch thread to start a new asynchronous fetch request.

get_current_locale

Returns the cached current system locale, or en-US by default.

set_default_accept_language

Step 12 of https://fetch.spec.whatwg.org/#concept-fetch

trim_http_whitespace

Normalize slice, as defined by the Fetch Spec.

Type Aliases

BoxedFetchCallback

CoreResourceThread

Handle to a resource thread