net_traits

Module blob_url_store

Structs

BlobBuf

Standalone blob buffer object

BlobResolver

BlobToken

A reference to a blob URL that will revoke the blob when dropped, unless the disown method is invoked.

BlobTokenCommunicator

TokenSerializationGuard

Guarantees that blob entries kept alive the contained token are not deallocated even if this token is serialized, dropped, and then later deserialized (possibly in a different thread).

UrlWithBlobClaim

This type upholds the variant that if the URL is a valid blob URL, then it has a token. Violating this invariant causes logic errors, but no unsafety.

Enums

BlobURLStoreError

Errors returned to Blob URL Store request

Functions

parse_blob_url

Parse URL as Blob URL scheme’s definition