Struct net_traits::blob_url_store::BlobURLStoreEntry

pub struct BlobURLStoreEntry {
    pub type_string: String,
    pub filename: Option<String>,
    pub size: u64,
    pub bytes: Vec<u8>,
}

Blob URL store entry, a packaged form of Blob DOM object

Fields
type_string: String

MIME type string

filename: Option<String>

Some filename if the backend of Blob is a file

size: u64

Size of content in bytes

bytes: Vec<u8>

Content of blob