net_traits::blob_url_store

Struct UrlWithBlobClaim

pub struct UrlWithBlobClaim {
    url: ServoUrl,
    token: Option<TokenSerializationGuard>,
}

This type upholds the variant that if the URL is a valid blob URL, then it has a token. Violating this invariant causes logic errors, but no unsafety.

Fields
url: ServoUrltoken: Option<TokenSerializationGuard>
Implementations

impl UrlWithBlobClaim

pub fn new(url: ServoUrl, token: Option<TokenSerializationGuard>) -> Self

pub fn token(&self) -> Option<&BlobToken>

pub fn blob_id(&self) -> Option<Uuid>

pub fn origin(&self) -> ImmutableOrigin

https://url.spec.whatwg.org/#concept-url-origin

pub fn for_url(url: ServoUrl) -> Result<Self, ServoUrl>

Constructs a UrlWithBlobClaim for URLs that are not blob URLs (Such URLs don’t need to claim anything).

Returns an Err containing the original URL if it’s a blob URL, so it can be reused without cloning.

pub fn from_url_without_having_claimed_blob(url: ServoUrl) -> Self

This method should only exist temporarily, and all callers should either claim the blob or guarantee that the URL is not a blob URL.

pub fn url(&self) -> ServoUrl