net_traits

Enum CoreResourceMsg

pub enum CoreResourceMsg {

    Fetch(RequestBuilder, FetchChannels),
    Cancel(Vec<RequestId>),
    FetchRedirect(RequestBuilder, ResponseInit, GenericCallback<FetchResponseMsg>),
    SetCookieForUrl(ServoUrl, Serde<Cookie<'static>>, CookieSource, Option<GenericSender<()>>),
    SetCookiesForUrl(ServoUrl, Vec<Serde<Cookie<'static>>>, CookieSource),
    SetCookieForUrlAsync(CookieStoreId, ServoUrl, Serde<Cookie<'static>>, CookieSource),
    GetCookieStringForUrl(ServoUrl, GenericSender<Option<String>>, CookieSource),
    GetCookiesForUrl(ServoUrl, GenericSender<Vec<Serde<Cookie<'static>>>>, CookieSource),
    EmbedderGetCookiesForUrl(CookieOperationId, ServoUrl, CookieSource),
    EmbedderSetCookieForUrl(CookieOperationId, ServoUrl, Serde<Cookie<'static>>, CookieSource),
    EmbedderClearCookies(CookieOperationId),
    EmbedderClearSessionCookies(CookieOperationId),
    GetCookieDataForUrlAsync(CookieStoreId, ServoUrl, Option<String>),
    GetAllCookieDataForUrlAsync(CookieStoreId, ServoUrl, Option<String>),
    DeleteCookiesForSites(Vec<String>, GenericSender<()>),
    DeleteCookies(Option<ServoUrl>, Option<GenericSender<()>>),
    DeleteSessionCookies(GenericSender<()>),
    DeleteCookie(ServoUrl, String),
    DeleteCookieAsync(CookieStoreId, ServoUrl, String),
    NewCookieListener(CookieStoreId, GenericCallback<CookieAsyncResponse>, ServoUrl),
    RemoveCookieListener(CookieStoreId),
    ListCookies(GenericSender<Vec<SiteDescriptor>>),
    GetHistoryState(HistoryStateId, GenericSender<Option<Vec<u8>>>),
    SetHistoryState(HistoryStateId, Vec<u8>),
    RemoveHistoryStates(Vec<HistoryStateId>),
    GetCacheEntries(GenericSender<Vec<CacheEntryDescriptor>>),
    ClearCache(Option<GenericSender<()>>),
    NetworkMediator(IpcSender<CustomResponseMediator>, ImmutableOrigin),
    ToFileManager(FileManagerThreadMsg),
    TotalSizeOfInFlightKeepAliveRecords(PipelineId, GenericSender<u64>),
    Exit(GenericOneshotSender<()>),
    CollectMemoryReport(ReportsChan),
    RevokeTokenForFile(BlobTokenRevocationRequest),
    RefreshTokenForFile(BlobTokenRefreshRequest),

}

Variants

Fetch(RequestBuilder, FetchChannels)

Cancel(Vec<RequestId>)

FetchRedirect(RequestBuilder, ResponseInit, GenericCallback<FetchResponseMsg>)

Initiate a fetch in response to processing a redirection

SetCookieForUrl(ServoUrl, Serde<Cookie<'static>>, CookieSource, Option<GenericSender<()>>)

Store a cookie for a given originating URL. If a sender is provided, the caller will block until the cookie is stored.

SetCookiesForUrl(ServoUrl, Vec<Serde<Cookie<'static>>>, CookieSource)

Store a set of cookies for a given originating URL

SetCookieForUrlAsync(CookieStoreId, ServoUrl, Serde<Cookie<'static>>, CookieSource)

GetCookieStringForUrl(ServoUrl, GenericSender<Option<String>>, CookieSource)

Retrieve the stored cookies as a header string for a given URL.

GetCookiesForUrl(ServoUrl, GenericSender<Vec<Serde<Cookie<'static>>>>, CookieSource)

Retrieve the stored cookies as a vector for the given URL. The response is sent via the provided sender.

EmbedderGetCookiesForUrl(CookieOperationId, ServoUrl, CookieSource)

Retrieve cookies for a URL for embedder. The response is sent via [NetToEmbedderMsg::EmbedderGetCookiesForUrlResponse].

EmbedderSetCookieForUrl(CookieOperationId, ServoUrl, Serde<Cookie<'static>>, CookieSource)

Set a cookie for a URL on behalf of the embedder. The response is sent via [NetToEmbedderMsg::EmbedderSetCookieForUrlResponse].

EmbedderClearCookies(CookieOperationId)

Clear all cookies on behalf of the embedder. The response is sent via NetToEmbedderMsg.

EmbedderClearSessionCookies(CookieOperationId)

Clear session cookies on behalf of the embedder. The response is sent via NetToEmbedderMsg.

GetCookieDataForUrlAsync(CookieStoreId, ServoUrl, Option<String>)

GetAllCookieDataForUrlAsync(CookieStoreId, ServoUrl, Option<String>)

DeleteCookiesForSites(Vec<String>, GenericSender<()>)

DeleteCookies(Option<ServoUrl>, Option<GenericSender<()>>)

This currently is used by unit tests and WebDriver only. When url is None, this clears cookies across all origins.

DeleteSessionCookies(GenericSender<()>)

Delete all session cookies (cookies without an expiry or max-age).

DeleteCookie(ServoUrl, String)

DeleteCookieAsync(CookieStoreId, ServoUrl, String)

NewCookieListener(CookieStoreId, GenericCallback<CookieAsyncResponse>, ServoUrl)

RemoveCookieListener(CookieStoreId)

ListCookies(GenericSender<Vec<SiteDescriptor>>)

GetHistoryState(HistoryStateId, GenericSender<Option<Vec<u8>>>)

Get a history state by a given history state id

SetHistoryState(HistoryStateId, Vec<u8>)

Set a history state for a given history state id

RemoveHistoryStates(Vec<HistoryStateId>)

Removes history states for the given ids

GetCacheEntries(GenericSender<Vec<CacheEntryDescriptor>>)

Gets a list of origin descriptors derived from entries in the cache

ClearCache(Option<GenericSender<()>>)

Clear the network cache.

NetworkMediator(IpcSender<CustomResponseMediator>, ImmutableOrigin)

Send the service worker network mediator for an origin to CoreResourceThread

ToFileManager(FileManagerThreadMsg)

Message forwarded to file manager’s handler

TotalSizeOfInFlightKeepAliveRecords(PipelineId, GenericSender<u64>)

Exit(GenericOneshotSender<()>)

Break the load handler loop, send a reply when done cleaning up local resources and exit

CollectMemoryReport(ReportsChan)

RevokeTokenForFile(BlobTokenRevocationRequest)

RefreshTokenForFile(BlobTokenRefreshRequest)