Enum net_traits::NetworkError

pub enum NetworkError {
    Internal(String),
    LoadCancelled,
    SslValidation(Url),
}

Network errors that have to be exported out of the loaders

Variants

Internal

Could be any of the internal errors, like unsupported scheme, connection errors, etc.

LoadCancelled

SslValidation

SSL validation error that has to be handled in the HTML parser