Enum net_traits::filemanager_thread::FileManagerThreadMsg

pub enum FileManagerThreadMsg {
    SelectFile(Vec<FilterPattern>, IpcSender<FileManagerResult<SelectedFile>>, FileOrigin, Option<String>),
    SelectFiles(Vec<FilterPattern>, IpcSender<FileManagerResult<Vec<SelectedFile>>>, FileOrigin, Option<Vec<String>>),
    ReadFile(IpcSender<FileManagerResult<ReadFileProgress>>, Uuid, bool, FileOrigin),
    PromoteMemory(BlobBuf, bool, IpcSender<Result<Uuid, BlobURLStoreError>>, FileOrigin),
    AddSlicedURLEntry(Uuid, RelativePos, IpcSender<Result<Uuid, BlobURLStoreError>>, FileOrigin),
    DecRef(Uuid, FileOrigin, IpcSender<Result<(), BlobURLStoreError>>),
    ActivateBlobURL(Uuid, IpcSender<Result<(), BlobURLStoreError>>, FileOrigin),
    RevokeBlobURL(Uuid, FileOrigin, IpcSender<Result<(), BlobURLStoreError>>),
}

Variants
SelectFile(Vec<FilterPattern>, IpcSender<FileManagerResult<SelectedFile>>, FileOrigin, Option<String>)

Select a single file. Last field is pre-selected file path for testing

SelectFiles(Vec<FilterPattern>, IpcSender<FileManagerResult<Vec<SelectedFile>>>, FileOrigin, Option<Vec<String>>)

Select multiple files. Last field is pre-selected file paths for testing

ReadFile(IpcSender<FileManagerResult<ReadFileProgress>>, Uuid, bool, FileOrigin)

Read FileID-indexed file in chunks, optionally check URL validity based on boolean flag

PromoteMemory(BlobBuf, bool, IpcSender<Result<Uuid, BlobURLStoreError>>, FileOrigin)

Add an entry as promoted memory-based blob and send back the associated FileID as part of a valid/invalid Blob URL depending on the boolean flag

AddSlicedURLEntry(Uuid, RelativePos, IpcSender<Result<Uuid, BlobURLStoreError>>, FileOrigin)

Add a sliced entry pointing to the parent FileID, and send back the associated FileID as part of a valid Blob URL

DecRef(Uuid, FileOrigin, IpcSender<Result<(), BlobURLStoreError>>)

Decrease reference count and send back the acknowledgement

ActivateBlobURL(Uuid, IpcSender<Result<(), BlobURLStoreError>>, FileOrigin)

Activate an internal FileID so it becomes valid as part of a Blob URL

RevokeBlobURL(Uuid, FileOrigin, IpcSender<Result<(), BlobURLStoreError>>)

Revoke Blob URL and send back the acknowledgement