Struct net_traits::filemanager_thread::RelativePos

pub struct RelativePos {
    pub start: i64,
    pub end: Option<i64>,
}

Relative slice positions of a sequence, whose semantic should be consistent with (start, end) parameters in https://w3c.github.io/FileAPI/#dfn-slice

Fields
start: i64

Relative to first byte if non-negative, relative to one past last byte if negative,

end: Option<i64>

Relative offset from first byte if Some(non-negative), relative to one past last byte if Some(negative), None if one past last byte

Methods

impl RelativePos

fn full_range() -> RelativePos

Full range from start to end

fn from_opts(start: Option<i64>, end: Option<i64>) -> RelativePos

Instantiate optional slice position parameters

fn slice_inner(&self, rel_pos: &RelativePos) -> RelativePos

Slice the inner sliced range by repositioning

fn to_abs_range(&self, size: usize) -> Range<usize>

Compute absolute range by giving the total size https://w3c.github.io/FileAPI/#slice-method-algo

fn from_abs_range(range: Range<usize>, size: usize) -> RelativePos

Inverse operation of to_abs_range