net_traits

Module image_cache

Structs

ImageLoadListener

This is optionally passed to the image cache when requesting and image, and returned to the specified event loop when the image load completes. It is typically used to trigger a reflow and/or repaint.

PendingImageId

The unique id for an image that has previously been requested.

PendingImageResponse

RasterizationCompleteResponse

VectorImage

Enums

Image

ImageCacheResponseMessage

ImageCacheResult

ImageOrMetadataAvailable

Indicating either entire image or just metadata availability

ImageResponse

The returned image.

Traits

FontResolver

An interface for resolving font families and styles for SVG images.

ImageCache

An ImageCache manages fetching and decoding images for a single Pipeline for its Document and all of its associated Workers.

ImageCacheFactory

A shared ImageCacheFactory is a per-process data structure used to create an ImageCache inside that process in any ScriptThread. This allows sharing the same font database (for SVGs) and also decoding thread pool among all ImageCaches in the same process.

Type Aliases

ImageCacheResponseCallback

VectorImageId