net_traits::image_cache

Trait FontResolver

pub trait FontResolver:
    Sync
    + Send
    + MallocSizeOf {
    // Required methods
    fn resolve(&self, font: &Font, database: &mut Arc<Database>) -> Option<ID>;
    fn resolve_fallback(
        &self,
        char: char,
        excluded: &[ID],
        database: &mut Arc<Database>,
    ) -> Option<ID>;
}

An interface for resolving font families and styles for SVG images.

Required Methods

fn resolve(&self, font: &Font, database: &mut Arc<Database>) -> Option<ID>

Attempt to resolve a font reference using the provided database of fonts. Adding new fonts to the database is allowed. Return an index into the database if the font resolves to an entry, otherwise return None.

fn resolve_fallback( &self, char: char, excluded: &[ID], database: &mut Arc<Database>, ) -> Option<ID>

Backup resolve. Find a font that can represent char and is not in excluded.