Enum net_traits::image_cache_thread::ImageCacheCommand

pub enum ImageCacheCommand {
    GetImageOrMetadataIfAvailable(ServoUrl, UsePlaceholder, CanRequestImages, IpcSender<Result<ImageOrMetadataAvailable, ImageState>>),
    AddListener(PendingImageId, ImageResponder),
    StoreDecodeImage(PendingImageId, FetchResponseMsg),
    Exit(IpcSender<()>),
}

Commands that the image cache understands.

Variants
GetImageOrMetadataIfAvailable(ServoUrl, UsePlaceholder, CanRequestImages, IpcSender<Result<ImageOrMetadataAvailable, ImageState>>)

Synchronously check the state of an image in the cache. If the image is in a loading state and but its metadata has been made available, it will be sent as a response.

AddListener(PendingImageId, ImageResponder)

Add a new listener for the given pending image.

StoreDecodeImage(PendingImageId, FetchResponseMsg)

Instruct the cache to store this data as a newly-complete network request and continue decoding the result into pixel data

Exit(IpcSender<()>)

Clients must wait for a response before shutting down the ResourceThread