Enum net_traits::image_cache_thread::ImageResponse

pub enum ImageResponse {
    Loaded(Arc<Image>),
    MetadataLoaded(ImageMetadata),
    PlaceholderLoaded(Arc<Image>),
    None,
}

The returned image.

Variants

Loaded

The requested image was loaded.

MetadataLoaded

The request image metadata was loaded.

PlaceholderLoaded

The requested image failed to load, so a placeholder was loaded instead.

None

Neither the requested image nor the placeholder could be loaded.