Struct net_traits::image_cache_thread::ImageCacheThread

pub struct ImageCacheThread {
    chan: IpcSender<ImageCacheCommand>,
}

The client side of the image cache thread. This can be safely cloned and passed to different threads.

Fields

chan

Methods

impl ImageCacheThread

The public API for the image cache thread.

fn new(chan: IpcSender<ImageCacheCommand>) -> ImageCacheThread

Construct a new image cache

fn request_image(&self, url: Url, result_chan: ImageCacheChan, responder: Option<ImageResponder>)

Asynchronously request an image. See ImageCacheCommand::RequestImage.

fn request_image_and_metadata(&self, url: Url, result_chan: ImageCacheChan, responder: Option<ImageResponder>)

Asynchronously request an image and metadata. See ImageCacheCommand::RequestImageAndMetadata

fn find_image(&self, url: Url, use_placeholder: UsePlaceholder) -> Result<Arc<Image>, ImageState>

Get the current state of an image. See ImageCacheCommand::GetImageIfAvailable.

fn find_image_or_metadata(&self, url: Url, use_placeholder: UsePlaceholder) -> Result<ImageOrMetadataAvailable, ImageState>

Get the current state of an image, returning its metadata if available. See ImageCacheCommand::GetImageOrMetadataIfAvailable.

fn store_complete_image_bytes(&self, url: Url, image_data: Vec<u8>)

Decode the given image bytes and cache the result for the given URL.

fn exit(&self)

Shutdown the image cache thread.