Struct net_traits::image_cache_thread::ImageResponder

pub struct ImageResponder {
    sender: IpcSender<ImageResponse>,
}

This is optionally passed to the image cache when requesting and image, and returned to the specified event loop when the image load completes. It is typically used to trigger a reflow and/or repaint.

Fields
sender: IpcSender<ImageResponse>
Methods

impl ImageResponder

fn new(sender: IpcSender<ImageResponse>) -> ImageResponder

fn respond(&self, response: ImageResponse)