Struct net_traits::response::Response

pub struct Response {

    pub response_type: ResponseType,
    pub termination_reason: Option<TerminationReason>,
    url: Option<ServoUrl>,
    pub url_list: Vec<ServoUrl>,
    pub status: Option<(StatusCode, String)>,
    pub raw_status: Option<(u16, Vec<u8>)>,
    pub headers: HeaderMap,
    pub body: Arc<Mutex<ResponseBody>>,
    pub cache_state: CacheState,
    pub https_state: HttpsState,
    pub referrer: Option<ServoUrl>,
    pub referrer_policy: Option<ReferrerPolicy>,
    pub cors_exposed_header_name_list: Vec<String>,
    pub location_url: Option<Result<ServoUrl, String>>,
    pub internal_response: Option<Box<Response>>,
    pub return_internal: bool,
    pub aborted: Arc<AtomicBool>,
    pub resource_timing: Arc<Mutex<ResourceFetchTiming>>,

}

A Response as defined by the Fetch spec

Fields
response_type: ResponseTypetermination_reason: Option<TerminationReason>url: Option<ServoUrl>url_list: Vec<ServoUrl>status: Option<(StatusCode, String)>

None can be considered a StatusCode of 0.

raw_status: Option<(u16, Vec<u8>)>headers: HeaderMapbody: Arc<Mutex<ResponseBody>>cache_state: CacheStatehttps_state: HttpsStatereferrer: Option<ServoUrl>referrer_policy: Option<ReferrerPolicy>cors_exposed_header_name_list: Vec<String>

CORS-exposed header-name list

location_url: Option<Result<ServoUrl, String>>

Location URL

internal_response: Option<Box<Response>>

Internal response, only used if the Response is a filtered response

return_internal: bool

whether or not to try to return the internal_response when asked for actual_response

aborted: Arc<AtomicBool>

https://fetch.spec.whatwg.org/#concept-response-aborted

resource_timing: Arc<Mutex<ResourceFetchTiming>>

track network metrics

Implementations

impl Response

pub fn new(url: ServoUrl, resource_timing: ResourceFetchTiming) -> Response

pub fn from_init( init: ResponseInit, resource_timing_type: ResourceTimingType ) -> Response

pub fn network_error(e: NetworkError) -> Response

pub fn url(&self) -> Option<&ServoUrl>

pub fn is_network_error(&self) -> bool

pub fn get_network_error(&self) -> Option<&NetworkError>

pub fn actual_response(&self) -> &Response

pub fn actual_response_mut(&mut self) -> &mut Response

pub fn to_actual(self) -> Response

pub fn get_resource_timing(&self) -> Arc<Mutex<ResourceFetchTiming>>

pub fn to_filtered(self, filter_type: ResponseType) -> Response

Convert to a filtered response, of type filter_type. Do not use with type Error or Default

pub fn metadata(&self) -> Result<FetchMetadata, NetworkError>