Enum net_traits::storage_thread::StorageThreadMsg

pub enum StorageThreadMsg {
    Length(IpcSender<usize>, Url, StorageType),
    Key(IpcSender<Option<String>>, Url, StorageType, u32),
    Keys(IpcSender<Vec<String>>, Url, StorageType),
    GetItem(IpcSender<Option<String>>, Url, StorageType, String),
    SetItem(IpcSender<Result<(bool, Option<String>), ()>>, Url, StorageType, String, String),
    RemoveItem(IpcSender<Option<String>>, Url, StorageType, String),
    Clear(IpcSender<bool>, Url, StorageType),
    Exit(IpcSender<()>),
}

Request operations on the storage data associated with a particular url

Variants
Length(IpcSender<usize>, Url, StorageType)

gets the number of key/value pairs present in the associated storage data

Key(IpcSender<Option<String>>, Url, StorageType, u32)

gets the name of the key at the specified index in the associated storage data

Keys(IpcSender<Vec<String>>, Url, StorageType)

Gets the available keys in the associated storage data

GetItem(IpcSender<Option<String>>, Url, StorageType, String)

gets the value associated with the given key in the associated storage data

SetItem(IpcSender<Result<(bool, Option<String>), ()>>, Url, StorageType, String, String)

sets the value of the given key in the associated storage data

RemoveItem(IpcSender<Option<String>>, Url, StorageType, String)

removes the key/value pair for the given key in the associated storage data

Clear(IpcSender<bool>, Url, StorageType)

clears the associated storage data by removing all the key/value pairs

Exit(IpcSender<()>)

send a reply when done cleaning up thread resources and then shut it down