Struct net_traits::LoadData

pub struct LoadData {
    pub url: Url,
    pub method: Method,
    pub headers: Headers,
    pub preserved_headers: Headers,
    pub data: Option<Vec<u8>>,
    pub cors: Option<ResourceCORSData>,
    pub pipeline_id: Option<PipelineId>,
    pub credentials_flag: bool,
    pub context: LoadContext,
    pub referrer_policy: Option<ReferrerPolicy>,
    pub referrer_url: Option<Url>,
}

Fields

url

method

headers

Headers that will apply to the initial request only

preserved_headers

Headers that will apply to the initial request and any redirects

data

cors

pipeline_id

credentials_flag

context

referrer_policy

The policy and referring URL for the originator of this request

referrer_url

Methods

impl LoadData

fn new(context: LoadContext, url: Url, id: Option<PipelineId>, referrer_policy: Option<ReferrerPolicy>, referrer_url: Option<Url>) -> LoadData