Struct net_traits::LoadResponse

pub struct LoadResponse {
    pub metadata: Metadata,
    pub progress_port: IpcReceiver<ProgressMsg>,
}

Message sent in response to Load. Contains metadata, and a port for receiving the data.

Even if loading fails immediately, we send one of these and the progress_port will provide the error.

Fields

metadata

Metadata, such as from HTTP headers.

progress_port

Port for reading data.