Struct net_traits::PendingAsyncLoad

pub struct PendingAsyncLoad {
    core_resource_thread: CoreResourceThread,
    url: Url,
    pipeline: Option<PipelineId>,
    guard: PendingLoadGuard,
    context: LoadContext,
    referrer_policy: Option<ReferrerPolicy>,
    referrer_url: Option<Url>,
    source: RequestSource,
}

Initialized but unsent request. Encapsulates everything necessary to instruct the resource thread to make a new request. The load method must be called before destruction or the thread will panic.

Fields

core_resource_thread

url

pipeline

guard

context

referrer_policy

referrer_url

source

Methods

impl PendingAsyncLoad

fn new(context: LoadContext, core_resource_thread: CoreResourceThread, url: Url, pipeline: Option<PipelineId>, referrer_policy: Option<ReferrerPolicy>, referrer_url: Option<Url>, source: RequestSource) -> PendingAsyncLoad

fn load_async(self, listener: AsyncResponseTarget)

Initiate the network request associated with this pending load, using the provided target.