net_traits

Trait FetchTaskTarget

pub trait FetchTaskTarget {
    // Required methods
    fn process_request_body(&mut self, request: &Request);
    fn process_response(&mut self, request: &Request, response: &Response);
    fn process_response_chunk(&mut self, request: &Request, chunk: Bytes);
    fn process_response_eof(&mut self, request: &Request, response: &Response);
    fn process_csp_violations(
        &mut self,
        request: &Request,
        violations: Vec<Violation>,
    );
    fn process_response_length_hint(
        &mut self,
        request_id: &Request,
        length: usize,
    );
}

Required Methods

fn process_request_body(&mut self, request: &Request)

https://fetch.spec.whatwg.org/#process-request-body

Fired when a chunk of the request body is transmitted

fn process_response(&mut self, request: &Request, response: &Response)

https://fetch.spec.whatwg.org/#process-response

Fired when headers are received

fn process_response_chunk(&mut self, request: &Request, chunk: Bytes)

Fired when a chunk of response content is received

fn process_response_eof(&mut self, request: &Request, response: &Response)

https://fetch.spec.whatwg.org/#process-response-end-of-file

Fired when the response is fully fetched

fn process_csp_violations( &mut self, request: &Request, violations: Vec<Violation>, )

fn process_response_length_hint(&mut self, request_id: &Request, length: usize)

Tell the listener that have a hint of how long the content is. This will be sent at most once.