Crate net_traits

Modules

blob_url_store

fetch

An implementation of the Fetch specification

filemanager_thread

http_status

image_cache

mime_classifier

policy_container

pub_domains

Implementation of public domain matching.

quality

request

response

The Response object resulting from a fetch operation

Structs

CacheEntryDescriptor

CookieAsyncResponse

CookieChange

CustomResponse

CustomResponseMediator

DebugVec

DiscardFetch

A fetch task that discards all data it’s sent, useful when speculatively prefetching data that we don’t need right now, but might need in the future.

FetchThread 🔒

A thread to handle fetches in a Servo process. This thread is responsible for listening for new fetch requests as well as updates on those operations and forwarding them to crossbeam channels.

Metadata

Metadata about a loaded resource, such as is obtained from HTTP headers.

ResourceCorsData

ResourceFetchTiming

ResourceThreads

SiteDescriptor

TlsSecurityInfo

Enums

CookieData

CookieSource

The creator of a given cookie

CoreResourceMsg

FetchChannels

IPC channels to communicate with the script thread about network or DOM events.

FetchMetadata

FetchResponseMsg

FilteredMetadata

IncludeSubdomains

LoadContext

A loading context, for context-specific sniffing, as defined in https://mimesniff.spec.whatwg.org/#context-specific-sniffing

MessageData

NetworkError

Network errors that have to be exported out of the loaders

RedirectEndValue

RedirectStartValue

ReferrerPolicy

Policies for providing a referrer header for a request

ResourceAttribute

ResourceTimeValue

ResourceTimingType

TlsSecurityState

ToFetchThreadMessage 🔒

WebSocketDomAction

WebSocketNetworkEvent

Constants

DOCUMENT_ACCEPT_HEADER_VALUE

https://fetch.spec.whatwg.org/#document-accept-header-value

Statics

FETCH_THREAD 🔒

PRIVILEGED_SECRET

Traits

AsyncRuntime

Handle to an async runtime, only used to shut it down for now.

FetchTaskTarget

Functions

cancel_async_fetch

Instruct the resource thread to cancel an existing request. Does nothing if the request has already completed or has not been fetched yet.

exit_fetch_thread

Send the exit message to the background thread, after which the caller can, and should, join on the thread.

fetch_async

Instruct the resource thread to make a new fetch request.

http_percent_encode

set_default_accept_language

Step 12 of https://fetch.spec.whatwg.org/#concept-fetch

start_fetch_thread

Start the fetch thread, and returns the join handle to the background thread.

trim_http_whitespace

Normalize slice, as defined by the Fetch Spec.

Type Aliases

BoxedFetchCallback

CoreResourceThread

Handle to a resource thread