Crate net_traits

Modules

blob_url_store

filemanager_thread

image

Image handling.

image_cache

pub_domains

Implementation of public domain matching.

quality

request

response

The Response object resulting from a fetch operation

storage_thread

Structs

CustomResponse

CustomResponseMediator

DiscardFetch

A fetch task that discards all data it’s sent, useful when speculatively prefetching data that we don’t need right now, but might need in the future.

Metadata

Metadata about a loaded resource, such as is obtained from HTTP headers.

PRIVILEGED_SECRET

ResourceCorsData

ResourceFetchTiming

ResourceThreads

WebrenderIpcSender

Enums

CookieSource

The creator of a given cookie

CoreResourceMsg

FetchChannels

IPC channels to communicate with the script thread about network or DOM events.

FetchMetadata

FetchResponseMsg

FilteredMetadata

IncludeSubdomains

LoadContext

A loading context, for context-specific sniffing, as defined in https://mimesniff.spec.whatwg.org/#context-specific-sniffing

MessageData

NetworkError

Network errors that have to be exported out of the loaders

RedirectEndValue

RedirectStartValue

ReferrerPolicy

Policies for providing a referrer header for a request

ResourceAttribute

ResourceTimeValue

ResourceTimingType

WebSocketDomAction

WebSocketNetworkEvent

WebrenderImageMsg

Traits

Action

FetchResponseListener

FetchTaskTarget

IpcSend

Abstraction of the ability to send a particular type of message, used by net_traits::ResourceThreads to ease the use its IpcSender sub-fields XXX: If this trait will be used more in future, some auto derive might be appealing

Functions

fetch_async

Instruct the resource thread to make a new request.

http_percent_encode

trim_http_whitespace

Normalize slice, as defined by the Fetch Spec.

Type Definitions

CoreResourceThread

Handle to a resource thread

IpcSendResult