Function net_traits::blob_url_store::get_blob_origin

pub fn get_blob_origin(url: &ServoUrl) -> FileOrigin

Given an URL, returning the Origin that a Blob created under this URL should have.

HACK(izgzhen): Not well-specified on spec, and it is a bit a hack both due to ambiguity of spec and that we have to serialization the Origin here.