Enum net_traits::CoreResourceMsg

pub enum CoreResourceMsg {
    Load(LoadData, LoadConsumer, Option<IpcSender<ResourceId>>),
    Fetch(RequestInit, IpcSender<FetchResponseMsg>),
    WebsocketConnect(WebSocketCommunicate, WebSocketConnectData),
    SetCookiesForUrl(Url, String, CookieSource),
    GetCookiesForUrl(Url, IpcSender<Option<String>>, CookieSource),
    Cancel(ResourceId),
    Synchronize(IpcSender<()>),
    Exit(IpcSender<()>),
}

Variants
Load(LoadData, LoadConsumer, Option<IpcSender<ResourceId>>)

Request the data associated with a particular URL

Fetch(RequestInit, IpcSender<FetchResponseMsg>)WebsocketConnect(WebSocketCommunicate, WebSocketConnectData)

Try to make a websocket connection to a URL.

SetCookiesForUrl(Url, String, CookieSource)

Store a set of cookies for a given originating URL

GetCookiesForUrl(Url, IpcSender<Option<String>>, CookieSource)

Retrieve the stored cookies for a given URL

Cancel(ResourceId)

Cancel a network request corresponding to a given ResourceId

Synchronize(IpcSender<()>)

Synchronization message solely for knowing the state of the ResourceChannelManager loop

Exit(IpcSender<()>)

Break the load handler loop, send a reply when done cleaning up local resources