Enum net_traits::CoreResourceMsg

pub enum CoreResourceMsg {
    Load(LoadData, LoadConsumer, Option<IpcSender<ResourceId>>),
    WebsocketConnect(WebSocketCommunicate, WebSocketConnectData),
    SetCookiesForUrl(Url, String, CookieSource),
    GetCookiesForUrl(Url, IpcSender<Option<String>>, CookieSource),
    Cancel(ResourceId),
    Synchronize(IpcSender<()>),
    Exit,
}

Variants

Load

Request the data associated with a particular URL

WebsocketConnect

Try to make a websocket connection to a URL.

SetCookiesForUrl

Store a set of cookies for a given originating URL

GetCookiesForUrl

Retrieve the stored cookies for a given URL

Cancel

Cancel a network request corresponding to a given ResourceId

Synchronize

Synchronization message solely for knowing the state of the ResourceChannelManager loop

Exit

Break the load handler loop and exit