Module net_traits::filemanager_thread

Structs

RelativePos

Relative slice positions of a sequence, whose semantic should be consistent with (start, end) parameters in https://w3c.github.io/FileAPI/#dfn-slice

SelectedFile

Response to file selection request

Enums

FileManagerThreadError

FileManagerThreadMsg

FileTokenCheck

A token modulating access to a file for a blob URL.

ReadFileProgress

Type Aliases

FileManagerResult

FileOrigin

File manager store entry’s origin