Module net_traits::image_cache

Structs

ImageResponder
This is optionally passed to the image cache when requesting and image, and returned to the specified event loop when the image load completes. It is typically used to trigger a reflow and/or repaint.

PendingImageId
The unique id for an image that has previously been requested.

PendingImageResponse

Enums

CorsStatus
Whether this response passed any CORS checks, and is thus safe to read from in cross-origin environments.

ImageCacheResult

ImageOrMetadataAvailable
Indicating either entire image or just metadata availability

ImageResponse
The returned image.

UsePlaceholder

Traits

ImageCache