Module net_traits::image_cache_thread

Structs

ImageCacheChan
Channel used by the image cache to send results.

ImageCacheResult
The result of an image cache command that is returned to the caller.

ImageCacheThread
The client side of the image cache thread. This can be safely cloned and passed to different threads.

ImageResponder
This is optionally passed to the image cache when requesting and image, and returned to the specified event loop when the image load completes. It is typically used to trigger a reflow and/or repaint.

Enums

ImageCacheCommand
Commands that the image cache understands.

ImageOrMetadataAvailable
Indicating either entire image or just metadata availability

ImageResponse
The returned image.

ImageState
The current state of an image in the cache.

UsePlaceholder

Constants

_IMPL_DESERIALIZE_FOR_ImageCacheChan

_IMPL_DESERIALIZE_FOR_ImageCacheCommand

_IMPL_DESERIALIZE_FOR_ImageCacheResult

_IMPL_DESERIALIZE_FOR_ImageCacheThread

_IMPL_DESERIALIZE_FOR_ImageOrMetadataAvailable

_IMPL_DESERIALIZE_FOR_ImageResponder

_IMPL_DESERIALIZE_FOR_ImageResponse

_IMPL_DESERIALIZE_FOR_ImageState

_IMPL_DESERIALIZE_FOR_UsePlaceholder

_IMPL_SERIALIZE_FOR_ImageCacheChan

_IMPL_SERIALIZE_FOR_ImageCacheCommand

_IMPL_SERIALIZE_FOR_ImageCacheResult

_IMPL_SERIALIZE_FOR_ImageCacheThread

_IMPL_SERIALIZE_FOR_ImageOrMetadataAvailable

_IMPL_SERIALIZE_FOR_ImageResponder

_IMPL_SERIALIZE_FOR_ImageResponse

_IMPL_SERIALIZE_FOR_ImageState

_IMPL_SERIALIZE_FOR_UsePlaceholder