Enum net_traits::image_cache_thread::ImageCacheCommand

pub enum ImageCacheCommand {
    RequestImage(Url, ImageCacheChan, Option<ImageResponder>),
    RequestImageAndMetadata(Url, ImageCacheChan, Option<ImageResponder>),
    GetImageIfAvailable(Url, UsePlaceholder, IpcSender<Result<Arc<Image>, ImageState>>),
    GetImageOrMetadataIfAvailable(Url, UsePlaceholder, IpcSender<Result<ImageOrMetadataAvailable, ImageState>>),
    StoreDecodeImage(Url, Vec<u8>),
    Exit(IpcSender<()>),
}

Commands that the image cache understands.

Variants

RequestImage

Request an image asynchronously from the cache. Supply a channel to receive the result, and optionally an image responder that is passed to the result channel.

RequestImageAndMetadata

Requests an image and a "metadata-ready" notification message asynchronously from the cache. The cache will make an effort to send metadata before the image is completely loaded. Supply a channel to receive the results, and optionally an image responder that is passed to the result channel.

GetImageIfAvailable

Synchronously check the state of an image in the cache. TODO(gw): Profile this on some real world sites and see if it's worth caching the results of this locally in each layout / paint thread.

GetImageOrMetadataIfAvailable

Synchronously check the state of an image in the cache. If the image is in a loading state and but its metadata has been made available, it will be sent as a response.

StoreDecodeImage

Instruct the cache to store this data as a newly-complete network request and continue decoding the result into pixel data

Exit

Clients must wait for a response before shutting down the ResourceThread