Module net_traits::request

Structs

Request

A Request as defined by the Fetch spec.

RequestBody

The net component’s view into https://fetch.spec.whatwg.org/#bodies

RequestBuilder

Enums

BodyChunkRequest

Messages used to implement https://fetch.spec.whatwg.org/#concept-request-transmit-body which are sent from net to script (with the exception of Done, which is sent from script to script).

BodyChunkResponse

Messages used to implement https://fetch.spec.whatwg.org/#concept-request-transmit-body which are sent from script to net.

BodySource

https://fetch.spec.whatwg.org/#concept-body-source

CacheMode

Cache mode

CorsSettings

CORS settings attribute

CredentialsMode

Request credentials mode

Destination

A request destination

Initiator

An initiator

Origin

A request origin

ParserMetadata

Parser Metadata

RedirectMode

Redirect mode

Referrer

A referer

RequestMode

A request mode

ResponseTainting

Response tainting

ServiceWorkersMode

Service-workers mode

Window

Window

Functions

convert_header_names_to_sorted_lowercase_set

https://fetch.spec.whatwg.org/#ref-for-convert-header-names-to-a-sorted-lowercase-set

get_cors_unsafe_header_names

https://fetch.spec.whatwg.org/#cors-unsafe-request-header-names

is_cors_non_wildcard_request_header_name

https://fetch.spec.whatwg.org/#cors-non-wildcard-request-header-name

is_cors_safelisted_language 🔒

is_cors_safelisted_method

https://fetch.spec.whatwg.org/#cors-safelisted-method

is_cors_safelisted_request_accept 🔒

is_cors_safelisted_request_content_type 🔒

is_cors_safelisted_request_header

is_cors_unsafe_request_header_byte 🔒