Module net_traits::response

The Response object resulting from a fetch operation

Structs

Response

A Response as defined by the Fetch spec

ResponseInit

Enums

CacheState

Cache state

HttpsState

Https state

ResponseBody

The response body can still be pushed to after fetch This provides a way to store unfinished response bodies

ResponseType

Response type

TerminationReason

Response termination reason