Struct net_traits::response::Response

pub struct Response {
    pub response_type: ResponseType,
    pub termination_reason: Option<TerminationReason>,
    url: Option<ServoUrl>,
    pub url_list: RefCell<Vec<ServoUrl>>,
    pub status: Option<StatusCode>,
    pub raw_status: Option<(u16, Vec<u8>)>,
    pub headers: Headers,
    pub body: Arc<Mutex<ResponseBody>>,
    pub cache_state: CacheState,
    pub https_state: HttpsState,
    pub referrer: Option<ServoUrl>,
    pub internal_response: Option<Box<Response>>,
    pub return_internal: Cell<bool>,
}

A Response as defined by the Fetch spec

Fields
response_type: ResponseType termination_reason: Option<TerminationReason> url: Option<ServoUrl> url_list: RefCell<Vec<ServoUrl>> status: Option<StatusCode>

None can be considered a StatusCode of 0.

raw_status: Option<(u16, Vec<u8>)> headers: Headers body: Arc<Mutex<ResponseBody>> cache_state: CacheState https_state: HttpsState referrer: Option<ServoUrl> internal_response: Option<Box<Response>>

Internal response, only used if the Response is a filtered response

return_internal: Cell<bool>

whether or not to try to return the internal_response when asked for actual_response

Methods

impl Response

fn new(url: ServoUrl) -> Response

fn network_error(e: NetworkError) -> Response

fn url(&self) -> Option<&ServoUrl>

fn is_network_error(&self) -> bool

fn get_network_error(&self) -> Option<&NetworkError>

fn actual_response(&self) -> &Response

fn to_actual(self) -> Response

fn to_filtered(self, filter_type: ResponseType) -> Response

Convert to a filtered response, of type filter_type. Do not use with type Error or Default

fn metadata(&self) -> Result<FetchMetadata, NetworkError>