Struct net_traits::response::Response

pub struct Response {
    pub response_type: ResponseType,
    pub termination_reason: Option<TerminationReason>,
    pub url: Option<Url>,
    pub url_list: RefCell<Vec<Url>>,
    pub status: Option<StatusCode>,
    pub headers: Headers,
    pub body: Arc<Mutex<ResponseBody>>,
    pub cache_state: CacheState,
    pub https_state: HttpsState,
    pub internal_response: Option<Box<Response>>,
    pub return_internal: Cell<bool>,
}

A Response as defined by the Fetch spec

Fields

response_type

termination_reason

url

url_list

status

None can be considered a StatusCode of 0.

headers

body

cache_state

https_state

internal_response

Internal response, only used if the Response is a filtered response

return_internal

whether or not to try to return the internal_response when asked for actual_response

Methods

impl Response

fn new() -> Response

fn network_error() -> Response

fn is_network_error(&self) -> bool

fn wait_until_done(&self)

fn actual_response(&self) -> &Response

fn to_actual(self) -> Response

fn to_filtered(self, filter_type: ResponseType) -> Response

Convert to a filtered response, of type filter_type. Do not use with type Error or Default