Trait net_traits::FetchTaskTarget

pub trait FetchTaskTarget {
    fn process_request_body(&mut self, request: &Request);
    fn process_request_eof(&mut self, request: &Request);
    fn process_response(&mut self, response: &Response);
    fn process_response_chunk(&mut self, chunk: Vec<u8>);
    fn process_response_eof(&mut self, response: &Response);
}

Required Methods

fn process_request_body(&mut self, request: &Request)

https://fetch.spec.whatwg.org/#process-request-body

Fired when a chunk of the request body is transmitted

fn process_request_eof(&mut self, request: &Request)

https://fetch.spec.whatwg.org/#process-request-end-of-file

Fired when the entire request finishes being transmitted

fn process_response(&mut self, response: &Response)

https://fetch.spec.whatwg.org/#process-response

Fired when headers are received

fn process_response_chunk(&mut self, chunk: Vec<u8>)

Fired when a chunk of response content is received

fn process_response_eof(&mut self, response: &Response)

https://fetch.spec.whatwg.org/#process-response-end-of-file

Fired when the response is fully fetched