net_traits

Trait FetchTaskTarget

pub trait FetchTaskTarget {
    // Required methods
    fn process_request_body(&mut self, request: &Request);
    fn process_request_eof(&mut self, request: &Request);
    fn process_response(&mut self, request: &Request, response: &Response);
    fn process_response_chunk(&mut self, request: &Request, chunk: Vec<u8>);
    fn process_response_eof(&mut self, request: &Request, response: &Response);
    fn process_csp_violations(
        &mut self,
        request: &Request,
        violations: Vec<Violation>,
    );
}

Required Methods

fn process_request_body(&mut self, request: &Request)

https://fetch.spec.whatwg.org/#process-request-body

Fired when a chunk of the request body is transmitted

fn process_request_eof(&mut self, request: &Request)

https://fetch.spec.whatwg.org/#process-request-end-of-file

Fired when the entire request finishes being transmitted

fn process_response(&mut self, request: &Request, response: &Response)

https://fetch.spec.whatwg.org/#process-response

Fired when headers are received

fn process_response_chunk(&mut self, request: &Request, chunk: Vec<u8>)

Fired when a chunk of response content is received

fn process_response_eof(&mut self, request: &Request, response: &Response)

https://fetch.spec.whatwg.org/#process-response-end-of-file

Fired when the response is fully fetched

fn process_csp_violations( &mut self, request: &Request, violations: Vec<Violation>, )